gpu: nvgpu: Cast int to u8 explicitly
In the Linux VM code the kind attribute is assigned to a u8 from an
int. The code is safe due to the checks in place, however, coverity
does not like this so explicitly cast to u8 before assigning kind
to kind_v.

Coverity ID: 2567919
Bug 200291879

Reviewed-on: https://git-master.nvidia.com/r/1548712
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/common/linux/vm.c
@@ -243,7 +243,7 @@ u64 nvgpu_vm_map(struct vm_gk20a *vm,
 		err = -EINVAL;
 		goto clean_up;
 	} else {
-		bfr.kind_v = kind;
+		bfr.kind_v = (u8)kind;
 	}
 	bfr.size = dmabuf->size;
 	sgl = bfr.sgt->sgl;
